changed type size in string ang vector to unsigned int

+++ OnlineLibrary/String.cpp
@@ -40,7 +40,7 @@ String::~String()
 	delete[] myString;
 }
 
-size_t String::getSize() const
+int String::getSize() const
 {
 	if (myString)
 	{
